+++ src/lib/jwt/Encoder.ts
@@ -1,5 +1,5 @@
 /**
- *  Copyright 2020 Angus.Fenying <fenying@litert.org>
+ *  Copyright 2021 Angus.Fenying <fenying@litert.org>
  *
  *  Licensed under the Apache License, Version 2.0 (the "License");
  *  you may not use this file except in compliance with the License.
@@ -14,100 +14,97 @@
  *  limitations under the License.
  */
 
-import * as Signs from "@litert/signatures";
-import * as Enc from "@litert/encodings";
+import * as Signs from '@litert/signatures';
+import * as Enc from '@litert/encodings';
 
-import * as C from "./Common";
-import * as E from "../Errors";
+import * as C from './Common';
+import * as E from '../Errors';
 
 interface IProfile {
 
-    jwa: C.JWA;
+    jwa: C.EJWA;
 
     name: string;
 
-    signer: Signs.ISigner<"base64url">;
+    signer: Signs.ISigner<'base64url'>;
+
+    predefinedHeaders: C.IRecurrsiveDict;
+
+    predefinedPayload: Partial<C.IJWTPayload>;
 }
 
 class JWTEncoder implements C.IJWTEncoder {
 
     private _profiles: Record<string, IProfile> = {};
 
-    public registerHMACProfile(
-        profileName: string,
-        algorithm: C.TValidHashAlgorithms,
-        key: string | Buffer
-    ): this {
+    public registerHMACProfile(opts: C.IHMACProfileOptions): this {
 
-        if (this._profiles[profileName]) {
+        if (this._profiles[opts.name]) {
 
-            throw new E.E_DUP_PROFILE({ metadata: { profileName} });
+            throw new E.E_DUP_PROFILE({ profileName: opts.name });
         }
 
-        this._profiles[profileName] = {
+        this._profiles[opts.name] = {
 
-            "jwa": (C.JWA as any)["HS" + algorithm.substr(3)],
-            "name": profileName,
-            "signer": Signs.HMAC.createSigner(
-                algorithm,
-                key,
-                "base64url"
-            )
+            'jwa': (C.EJWA as any)['HS' + opts.algorithm.substr(3)],
+            'name': opts.name,
+            'signer': Signs.HMAC.createSigner(
+                opts.algorithm,
+                opts.key,
+                'base64url'
+            ),
+            'predefinedHeaders': opts.predefinedHeaders ?? {},
+            'predefinedPayload': opts.predefinedPayload ?? {}
         };
 
         return this;
     }
 
-    public registerRSAProfile(
-        profileName: string,
-        algorithm: C.TValidHashAlgorithms,
-        key: C.IKeyPair,
-        pssmgf1Padding?: boolean
-    ): this {
+    public registerRSAProfile(opts: C.IRSAProfileOptions): this {
 
-        if (this._profiles[profileName]) {
+        if (this._profiles[opts.name]) {
 
-            throw new E.E_DUP_PROFILE({ metadata: { profileName} });
+            throw new E.E_DUP_PROFILE({ profileName: opts.name });
         }
 
-        this._profiles[profileName] = {
+        this._profiles[opts.name] = {
 
-            "jwa": (C.JWA as any)[pssmgf1Padding ? "PS" : "RS" + algorithm.substr(3)],
-            "name": profileName,
-            "signer": Signs.RSA.createSigner(
-                algorithm,
-                key.public,
-                key.private,
-                pssmgf1Padding ? Signs.ERSAPadding.PSS_MGF1 : Signs.ERSAPadding.PKCS1_V1_5,
-                "base64url"
-            )
+            'jwa': (C.EJWA as any)[(opts.pssmgf1Padding ? 'PS' : 'RS') + opts.algorithm.substr(3)],
+            'name': opts.name,
+            'signer': Signs.RSA.createSigner(
+                opts.algorithm,
+                opts.key.public,
+                opts.key.private,
+                opts.pssmgf1Padding ? Signs.ERSAPadding.PSS_MGF1 : Signs.ERSAPadding.PKCS1_V1_5,
+                'base64url'
+            ),
+            'predefinedHeaders': opts.predefinedHeaders ?? {},
+            'predefinedPayload': opts.predefinedPayload ?? {}
         };
 
         return this;
     }
 
-    public registerECDSAProfile(
-        profileName: string,
-        algorithm: C.TValidHashAlgorithms,
-        key: C.IKeyPair
-    ): this {
+    public registerECDSAProfile(opts: C.IECDSAProfileOptions): this {
 
-        if (this._profiles[profileName]) {
+        if (this._profiles[opts.name]) {
 
-            throw new E.E_DUP_PROFILE({ metadata: { profileName} });
+            throw new E.E_DUP_PROFILE({ profileName: opts.name });
         }
 
-        this._profiles[profileName] = {
+        this._profiles[opts.name] = {
 
-            "name": profileName,
-            "jwa": (C.JWA as any)["ES" + algorithm.substr(3)],
-            "signer": Signs.ECDSA.createSigner(
-                algorithm,
-                key.public,
-                key.private,
+            'jwa': (C.EJWA as any)['ES' + opts.algorithm.substr(3)],
+            'name': opts.name,
+            'signer': Signs.ECDSA.createSigner(
+                opts.algorithm,
+                opts.key.public,
+                opts.key.private,
                 Signs.EECDSAFormat.IEEE_P1363,
-                "base64url"
-            )
+                'base64url'
+            ),
+            'predefinedHeaders': opts.predefinedHeaders ?? {},
+            'predefinedPayload': opts.predefinedPayload ?? {}
         };
 
         return this;
@@ -120,7 +117,7 @@ class JWTEncoder implements C.IJWTEncoder {
         return this;
     }
 
-    public create(profileName: string, payload: C.IJWTPayload): string {
+    public create(profileName: string, payload: C.IJWTPayload, headers?: C.IRecurrsiveDict): string {
 
         const profile = this._profiles[profileName];
 
@@ -130,9 +127,14 @@ class JWTEncoder implements C.IJWTEncoder {
         }
 
         const body = Enc.stringToBase64Url(JSON.stringify({
-            "typ": "JWT",
-            "alg": C.JWA[profile.jwa]
-        })) + "." + Enc.stringToBase64Url(JSON.stringify(payload));
+            'typ': 'JWT',
+            'alg': C.EJWA[profile.jwa],
+            ...profile.predefinedHeaders,
+            ...headers
+        })) + '.' + Enc.stringToBase64Url(JSON.stringify({
+            ...profile.predefinedPayload,
+            ...payload
+        }));
 
         return `${body}.${profile.signer.sign(body)}`;
     }
@@ -146,7 +148,7 @@ class JWTEncoder implements C.IJWTEncoder {
             throw new E.E_PROFILE_NOT_FOUND({ metadata: { profileName} });
         }
 
-        return C.JWA[profile.jwa] === algo && profile.signer.verify(
+        return C.EJWA[profile.jwa] === algo && profile.signer.verify(
             body,
             signature
         );
@@ -154,14 +156,14 @@ class JWTEncoder implements C.IJWTEncoder {
 
     public verify(jwt: string, profileName?: string): C.IJWT {
 
-        const segs = jwt.split(".");
+        const segs = jwt.split('.');
 
         if (segs.length !== 3) {
 
             throw new E.E_MALFORMED_JWT();
         }
 
-        let header!: { "typ": string; "alg": string; };
+        let header!: { 'typ': string; 'alg': string; };
 
         let payload!: C.IJWTPayload;
 
@@ -177,7 +179,7 @@ class JWTEncoder implements C.IJWTEncoder {
             throw new E.E_MALFORMED_JWT({ metadata: { origin: e } });
         }
 
-        if (typeof header !== "object" || header.typ !== "JWT" || typeof header.alg !== "string") {
+        if (typeof header !== 'object' || header.typ !== 'JWT' || typeof header.alg !== 'string') {
 
             throw new E.E_MALFORMED_JWT();
         }
@@ -204,9 +206,9 @@ class JWTEncoder implements C.IJWTEncoder {
 
         return {
             header: {
-                type: "JWT",
-                algorithm: C.JWA[header.alg as any] as any,
-            },
+                ...header,
+                alg: C.EJWA[header.alg as any] as any,
+            } as any,
             payload,
             signature: {
                 value: segs[2],