+++ CHANGES.md
@@ -1,5 +1,11 @@
 # Changes Logs
 
+## v0.3.0
+
+- feat(jwt): added pre-defined JWT header in profile.
+- feat(jwt): added pre-defined JWT payload in profile.
+- feat(jwt): refactored the JWT helper APIs.
+
 ## v0.2.1
 
 - Updated dependencies and fixed the error definitions.